+++ examples/webpack-plugin-demo/entry/src/ohosTest/java/org/originjs/openharmony/webpackplugindemo/ExampleOhosTest.java
@@ -1,0 +1,14 @@
+package org.originjs.openharmony.webpackplugindemo;
+
+import ohos.aafwk.ability.delegation.AbilityDelegatorRegistry;
+import org.junit.Test;
+
+import static org.junit.Assert.assertEquals;
+
+public class ExampleOhosTest {
+    @Test
+    public void testBundleName() {
+        final String actualBundleName = AbilityDelegatorRegistry.getArguments().getTestBundleName();
+        assertEquals("org.originjs.openharmony.webpackplugindemo", actualBundleName);
+    }
+}