+++ packages/webpack-plugin/CHANGELOG.md
@@ -1,3 +1,13 @@
+# 0.6.0 (2022-02-25)
+
+
+### Features
+
+* support fs in webpack plugin ([2888fc5](https://github.com/originjs/openharmony-polyfill/commit/2888fc5c41699c9af811a2f4161993b8f99b8b36))
+
+* support crypto in webpack plugin
+
+
 # 0.1.0 (2022-02-19)
 
 ### Features